Make TS module resolution method explicit
Make the TS module resolution method explicit. The TypeScript
documentation at https://www.typescriptlang.org/docs/handbook/module-resolution.html
claims that node-style module resolution is the default if `module` is
set to `commonjs` but this doesn't seem to be the case? Or maybe we are
accidentally running on an old version of NodeJS in the Action in which
case the module resolution method would be classic.